+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/TigerTestEnvMgrApplication.java
@@ -31,5 +31,4 @@ public class TigerTestEnvMgrApplication implements ServletContextListener {
         log.info(Ansi.colorize("Tiger standalone test environment manager UP!", RbelAnsiColors.GREEN_BOLD));
         return envMgr;
     }
-
 }