+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/TigerTestEnvShutdownManager.java
@@ -1,0 +1,2 @@
+package de.gematik.test.tiger.testenvmgr;public class TigerTestEnvShutdownManager {
+}